+++ megamek/src/megamek/common/Compute.java
@@ -1527,13 +1527,13 @@ public class Compute {
 
             if ((wt.getAmmoType() == AmmoType.T_AC_ULTRA)
                     || (wt.getAmmoType() == AmmoType.T_AC_ROTARY)) {
-                fDamage = fHits * (float) wt.getDamage();
+                fDamage = fHits * wt.getDamage();
             }
         } else {
 
             // Direct fire weapons (and LBX slug rounds) just do a single shot
             // so they don't use the missile hits table
-            fDamage = (float) wt.getDamage();
+            fDamage = wt.getDamage();
 
             // Infantry follow some special rules, but do fixed amounts of
             // damage
@@ -1552,7 +1552,7 @@ public class Compute {
                         if ((g.getEntity(waa.getTargetId())
                                 .getSwarmAttackerId() == Entity.NONE)
                                 && (g.getEntity(waa.getTargetId()) instanceof Mech)) {
-                            fDamage = 1.5f * (float) inf_attacker
+                            fDamage = 1.5f * inf_attacker
                                     .getDamage(inf_attacker
                                             .getShootingStrength());
                         }
@@ -1565,7 +1565,7 @@ public class Compute {
                     else {
                         // conventional weapons; field guns should be handled
                         // under the standard weapons section
-                        fDamage = (float) inf_attacker.getDamage(inf_attacker
+                        fDamage = inf_attacker.getDamage(inf_attacker
                                 .getShootingStrength());
                     }
